restructure tests to have a shared utils library that is linked to
