+++ tests/test_construct_problem.cpp
@@ -10,7 +10,8 @@
  */
 
 #include <CORA/CORA_problem.h>
-#include <CORA/CORA_test_utils.h>
+
+#include <test_utils.h>
 
 #include <string>
