+++ tests/test_cora.cpp
@@ -1,6 +1,6 @@
 #include <CORA/CORA.h>
-#include <CORA/CORA_test_utils.h>
 #include <CORA/pyfg_text_parser.h>
+#include <test_utils.h>
 
 #include <filesystem>
 #include <string>
