+++ tests/test_parse_pyfg.cpp
@@ -2,8 +2,9 @@
 // Created by Tim Magoun on 10/31/23.
 //
 
-#include <CORA/CORA_test_utils.h>
 #include <CORA/pyfg_text_parser.h>
+
+#include <test_utils.h>
 
 #include <filesystem>
 #include <string>
